Organize assets, constants folder created, links fixed, small fixes
