+++ src/components/SocialIcon/SocialIcon.component.tsx
@@ -1,16 +1,17 @@
-import fbIcon from '@/assets/fb.svg';
-import inst from '@/assets/insta.svg';
-import linkedin from '@/assets/linkedin.svg';
-import styles from '@/components/footer/footer.module.css';
+import fbIcon from '@/assets/footerImages/fb.svg';
+import inst from '@/assets/footerImages/insta.svg';
+import linkedin from '@/assets/footerImages/linkedin.svg';
+import styles from '@/components/footer/Footer.module.css';
 import FooterIcon from '@/components/footer/FooterIcon.tsx';
+import SocialLinks from '@/constants/SocialLinks.tsx';
 
 function SocialIcon() {
     return (
         <>
             <div className={styles.socialIcons}>
-                <FooterIcon iconPath={fbIcon} link="https://www.facebook.com/profile.php?id=100086849614305" />
-                <FooterIcon iconPath={linkedin} link="https://linkedin.com/" />
-                <FooterIcon iconPath={inst} link="https://www.instagram.com/" />
+                <FooterIcon iconPath={fbIcon} link={SocialLinks.facebook} />
+                <FooterIcon iconPath={linkedin} link={SocialLinks.linkedin} />
+                <FooterIcon iconPath={inst} link={SocialLinks.instagram} />
             </div>
         </>
     );
