+++ src/components/about/AboutPage.component.tsx
@@ -1,4 +1,5 @@
-import aboutMe from '@/assets/about_me.svg';
+import aboutMe from '@/assets/aboutImages/aboutMe.svg';
+import SocialLinks from '@/constants/SocialLinks.tsx';
 
 import styles from './AboutPage.module.css';
 
@@ -17,7 +18,7 @@ function About() {
                                 </p>
                                 <p className={styles.aboutParagraph}>
                                     This page was developed during the course{' '}
-                                    <a className={styles.link} href="https://www.mastersacademy.education/frontend-react-it">
+                                    <a className={styles.link} href={SocialLinks.masters}>
                                         &apos;Intro to React&apos;
                                     </a>{' '}
                                     from Masters Academy in 2024.
