+++ src/components/footer/Footer.component.tsx
@@ -1,6 +1,7 @@
 import SocialIcon from '@/components/SocialIcon/SocialIcon.component.tsx';
+import SocialLinks from '@/constants/SocialLinks.tsx';
 
-import styles from './footer.module.css';
+import styles from './Footer.module.css';
 function FooterComponent() {
     return (
         <>
@@ -11,7 +12,7 @@ function FooterComponent() {
                         <div className={styles.footerText}>
                             <p>
                                 Made with 💗 on course{' '}
-                                <a className={styles.linkFooter} href="https://www.mastersacademy.education/frontend-for-beginners-it">
+                                <a className={styles.linkFooter} href={SocialLinks.masters}>
                                     &apos;Intro to React&apos; from Masters Academy in 2024
                                 </a>{' '}
                                 by Dmytro Shevchenko
