+++ src/components/footer/FooterIcon.tsx
@@ -1,4 +1,4 @@
-import styles from './footer.module.css';
+import styles from './Footer.module.css';
 
 interface FooterIconProps {
     link: string;
