+++ src/constants/SocialLinks.tsx
@@ -1,0 +1,8 @@
+const SocialLinks = {
+    masters: 'https://www.mastersacademy.education/frontend-react-it',
+    facebook: 'https://www.facebook.com/profile.php?id=100086849614305',
+    linkedin: 'https://linkedin.com/',
+    instagram: 'https://www.instagram.com/',
+};
+
+export default SocialLinks;
