+++ src/interfaces/Category.tsx
@@ -1,0 +1,7 @@
+export interface Category {
+    id: number;
+    name: string;
+    image: string;
+    createdAt: string;
+    updatedAt: string;
+}
